feat: support write partition table in grpc service
